Feat: 모달창 쿠키 데이터로 업데이트
