+++ style.css
@@ -171,7 +171,7 @@ hr {
 }
 .wish-container{
     max-height: 25rem;
-    max-width: 22.5rem;
+    width: 22.5rem;
     overflow-y: auto;
     display: flex;
     flex-direction: column;
